Machine Learning/requirements.txt
Flask==3.0.3
pandas==2.2.2
gunicorn==22.0.0
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
matplotlib==3.9.0
